refactor: move lambda types to separate file
