+++ lambda/index.ts
@@ -44,30 +44,7 @@ import {
   Logger,
 } from 'aws-cloudformation-custom-resource';
 import * as forge from 'node-forge';
-
-export enum PublicKeyFormat {
-  /* eslint-disable @typescript-eslint/naming-convention */
-  OPENSSH = 'OPENSSH',
-  PEM = 'PEM',
-  /* eslint-enable @typescript-eslint/naming-convention */
-}
-
-export interface ResourceProperties {
-  /* eslint-disable @typescript-eslint/naming-convention */
-  Name: string;
-  StorePublicKey?: 'true' | 'false'; // props passed via lambda always are of type string
-  ExposePublicKey?: 'true' | 'false';
-  PublicKey: string;
-  SecretPrefix: string;
-  Description: string;
-  KmsPrivate: string;
-  KmsPublic: string;
-  PublicKeyFormat: PublicKeyFormat;
-  RemoveKeySecretsAfterDays: number;
-  StackName: string;
-  Tags: Record<string, string>;
-  /* eslint-enable @typescript-eslint/naming-convention */
-}
+import { PublicKeyFormat, ResourceProperties } from './types';
 
 const ec2Client = new EC2Client({});
 const secretsManagerClient = new SecretsManagerClient({});
